Refactor hierarchy of the proxy subpackages
