+++ src/test/java/com/gfl/client/service/proxy/ProxySourceServiceFileTest.java
@@ -1,5 +1,7 @@
 package com.gfl.client.service.proxy;
 
+import com.gfl.client.service.proxy.source.ProxySourceService;
+import com.gfl.client.service.proxy.source.ProxySourceServiceFile;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
